fix work with sessions

+++ Cimas/Models/From/AddFilmModel.cs
@@ -6,7 +6,7 @@ namespace Cimas.Models.From
     {
         [Required]
         public string Name { get; set; }
-        [Range(1, int.MaxValue, ErrorMessage = "Duration can't be less than 1")]
+        [Range(1, double.MaxValue, ErrorMessage = "Duration can't be less than 1")]
         public double Duration { get; set; }
     }
 }
